split dictionary consts from manifests
