+++ src/packages/dictionary/collection/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_COLLECTION_ALIAS = 'Umb.Collection.Dictionary';
