+++ src/packages/dictionary/collection/index.ts
@@ -1,2 +1,2 @@
 export { UmbDictionaryCollectionRepository } from './repository/index.js';
-export { UMB_DICTIONARY_COLLECTION_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_COLLECTION_ALIAS } from './constants.js';
