+++ src/packages/dictionary/collection/manifests.ts
@@ -2,9 +2,8 @@ import { UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS } from './repository/index.j
 import { manifests as collectionRepositoryManifests } from './repository/manifests.js';
 import { manifests as collectionViewManifests } from './views/manifests.js';
 import { manifests as collectionActionManifests } from './action/manifests.js';
+import { UMB_DICTIONARY_COLLECTION_ALIAS } from './constants.js';
 import type { ManifestCollection, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_COLLECTION_ALIAS = 'Umb.Collection.Dictionary';
 
 const collectionManifest: ManifestCollection = {
 	type: 'collection',
