+++ src/packages/dictionary/collection/repository/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Collection';
