+++ src/packages/dictionary/collection/repository/index.ts
@@ -1,2 +1,2 @@
-export { UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS } from './constants.js';
 export { UmbDictionaryCollectionRepository } from './dictionary-collection.repository.js';
