+++ src/packages/dictionary/collection/repository/manifests.ts
@@ -1,6 +1,5 @@
+import { UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_COLLECTION_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Collection';
 
 const repository: ManifestRepository = {
 	type: 'repository',
