+++ src/packages/dictionary/collection/views/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_TABLE_COLLECTION_VIEW_ALIAS = 'Umb.CollectionView.Dictionary.Table';
