+++ src/packages/dictionary/collection/views/index.ts
@@ -1,1 +1,1 @@
-export { UMB_DICTIONARY_TABLE_COLLECTION_VIEW_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_TABLE_COLLECTION_VIEW_ALIAS } from './constants.js';
