+++ src/packages/dictionary/menu/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_MENU_ALIAS = 'Umb.Menu.Dictionary';
