+++ src/packages/dictionary/menu/index.ts
@@ -1,0 +1,1 @@
+export { UMB_DICTIONARY_MENU_ALIAS } from './constants.js';
