+++ src/packages/dictionary/menu/manifests.ts
@@ -1,8 +1,7 @@
 import { UMB_DICTIONARY_ENTITY_TYPE } from '../entity.js';
 import { UMB_DICTIONARY_TREE_ALIAS } from '../tree/index.js';
+import { UMB_DICTIONARY_MENU_ALIAS } from './constants.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_MENU_ALIAS = 'Umb.Menu.Dictionary';
 
 export const manifests: Array<ManifestTypes> = [
 	{
