+++ src/packages/dictionary/repository/detail/index.ts
@@ -1,2 +1,2 @@
 export { UmbDictionaryDetailRepository } from './dictionary-detail.repository.js';
-export { UMB_DICTIONARY_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_DETAIL_REPOSITORY_ALIAS } from './constants.js';
