+++ src/packages/dictionary/repository/export/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_EXPORT_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Export';
