+++ src/packages/dictionary/repository/export/index.ts
@@ -1,2 +1,2 @@
 export { UmbDictionaryExportRepository } from './dictionary-export.repository.js';
-export { UMB_DICTIONARY_EXPORT_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_EXPORT_REPOSITORY_ALIAS } from './constants.js';
