+++ src/packages/dictionary/repository/export/manifests.ts
@@ -1,7 +1,6 @@
+import { UMB_DICTIONARY_EXPORT_REPOSITORY_ALIAS } from './constants.js';
 import { UmbDictionaryExportRepository } from './dictionary-export.repository.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_EXPORT_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Export';
 
 const repository: ManifestRepository = {
 	type: 'repository',
