+++ src/packages/dictionary/repository/import/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_IMPORT_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Import';
