+++ src/packages/dictionary/repository/import/index.ts
@@ -1,2 +1,2 @@
 export { UmbDictionaryImportRepository } from './dictionary-import.repository.js';
-export { UMB_DICTIONARY_IMPORT_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_IMPORT_REPOSITORY_ALIAS } from './constants.js';
