+++ src/packages/dictionary/repository/import/manifests.ts
@@ -1,7 +1,6 @@
+import { UMB_DICTIONARY_IMPORT_REPOSITORY_ALIAS } from './constants.js';
 import { UmbDictionaryImportRepository } from './dictionary-import.repository.js';
 import type { ManifestRepository, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_IMPORT_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Import';
 
 const repository: ManifestRepository = {
 	type: 'repository',
