+++ src/packages/dictionary/repository/item/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_DICTIONARY_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Item';
+export const UMB_DICTIONARY_STORE_ALIAS = 'Umb.Store.Dictionary.Item';
