+++ src/packages/dictionary/repository/item/index.ts
@@ -1,3 +1,3 @@
 export { UmbDictionaryItemRepository } from './dictionary-item.repository.js';
-export { UMB_DICTIONARY_ITEM_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_DICTIONARY_ITEM_REPOSITORY_ALIAS } from './constants.js';
 export type { UmbDictionaryItemModel } from './types.js';
