+++ src/packages/dictionary/repository/item/manifests.ts
@@ -1,9 +1,7 @@
 import { UmbDictionaryItemStore } from './dictionary-item.store.js';
 import { UmbDictionaryItemRepository } from './dictionary-item.repository.js';
+import { UMB_DICTIONARY_ITEM_REPOSITORY_ALIAS, UMB_DICTIONARY_STORE_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestItemStore, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_DICTIONARY_ITEM_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Item';
-export const UMB_DICTIONARY_STORE_ALIAS = 'Umb.Store.Dictionary.Item';
 
 const itemRepository: ManifestRepository = {
 	type: 'repository',
