+++ src/packages/dictionary/section/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_DICTIONARY_SECTION_ALIAS = 'Umb.Section.Translation';
