+++ src/packages/dictionary/section/index.ts
@@ -1,1 +1,1 @@
-export { UMB_DICTIONARY_SECTION_ALIAS } from '../section/manifests.js';
+export { UMB_DICTIONARY_SECTION_ALIAS } from './constants.js';
