+++ src/packages/dictionary/tree/constants.ts
@@ -1,0 +1,3 @@
+export const UMB_DICTIONARY_TREE_REPOSITORY_ALIAS = 'Umb.Repository.Dictionary.Tree';
+export const UMB_DICTIONARY_TREE_STORE_ALIAS = 'Umb.Store.Dictionary.Tree';
+export const UMB_DICTIONARY_TREE_ALIAS = 'Umb.Tree.Dictionary';
