+++ src/packages/dictionary/tree/index.ts
@@ -3,7 +3,7 @@ export {
 	UMB_DICTIONARY_TREE_REPOSITORY_ALIAS,
 	UMB_DICTIONARY_TREE_STORE_ALIAS,
 	UMB_DICTIONARY_TREE_ALIAS,
-} from './manifests.js';
+} from './constants.js';
 export { UMB_DICTIONARY_TREE_STORE_CONTEXT } from './dictionary-tree.store.js';
 export { type UmbDictionaryTreeStore } from './dictionary-tree.store.js';
 export * from './types.js';
